Ensure Initialize can run multiple times
My previous change in #13 introduced the bug that, if the
schema_migrations table exists but was empty, the Initialize function
would be called and then fail attempting to create the table.

Instead, I've pulled the check for the table's existence into Initialize
itself, so that it can simply opt-out of creating it.